requirements.txt
black==21.12b0
celery==5.2.3
coverage==6.2
Django>=3.2,<3.3
django-celery-beat==2.2.1
django-extensions==3.1.5
django-leaflet==0.28.2
django-redis==5.2.0
# django-zen-queries==2.1.0
flake8==4.0.1
geopandas==0.10.2
ipython==7.31.0
isort==5.10.1
jupyter==1.0.0
lxml==4.7.1
notebook==6.4.6
# pandas==1.3.5
psycopg2-binary==2.9.3
pytest-django==4.5.2
pytest-cov==3.0.0
uvicorn==0.16.0
wagtail>=2.15,<2.16
Werkzeug==2.0.2  # needed for runserver_plus
